+++ examples/node/basic_ts_esm/example.ts
@@ -1,22 +1,23 @@
 import * as Dotenv from "dotenv";
-import { AllowArbitraryParams, config, getJson, GoogleParameters } from "serpapi";
+import { config, EngineParameters, getJson } from "serpapi";
 
 Dotenv.config();
 const apiKey = process.env.API_KEY;
 
 const params = {
+  engine: "google",
   q: "Coffee",
   api_key: apiKey,
-} satisfies AllowArbitraryParams<GoogleParameters>;
+} satisfies EngineParameters<"google">;
 
 // Show result as JSON (async/await)
-const response1 = await getJson("google", params);
+const response1 = await getJson(params);
 console.log(response1["organic_results"]);
 
 // Show result as JSON (callback)
-getJson("google", params, (json) => console.log(json["organic_results"]));
+getJson(params, (json) => console.log(json["organic_results"]));
 
 // Use global config
 config.api_key = apiKey;
-const response2 = await getJson("google", { q: "Coffee" });
+const response2 = await getJson({ engine: "google", q: "Coffee" });
 console.log(response2["organic_results"]);