+++ examples/node/pagination_js_commonjs/example.js
@@ -8,12 +8,13 @@ const extractLinks = (results) => results.map((r) => r.link);
 
 const run = async () => {
   const params = {
+    engine: "google",
     q: "Coffee",
     api_key: apiKey,
   };
 
   // Pagination (async/await)
-  let page1 = await getJson("google", params);
+  let page1 = await getJson(params);
   console.log(
     "First page links",
     extractLinks(page1.organic_results),
@@ -25,7 +26,7 @@ const run = async () => {
   );
 
   // Pagination (callback)
-  getJson("google", params, (page1) => {
+  getJson(params, (page1) => {
     console.log(
       "First page links",
       extractLinks(page1.organic_results),
@@ -40,7 +41,7 @@ const run = async () => {
 
   // Use global config
   config.api_key = apiKey;
-  page1 = await getJson("google", { q: "Coffee" });
+  page1 = await getJson({ engine: "google", q: "Coffee" });
   page2 = await page1.next?.();
   console.log(
     "Second page links",
@@ -49,7 +50,7 @@ const run = async () => {
 
   // Pagination loop (async/await)
   let links = [];
-  let page = await getJson("google", { q: "Coffee" });
+  let page = await getJson({ engine: "google", q: "Coffee" });
   while (page) {
     links.push(...extractLinks(page.organic_results));
     if (links.length >= 30) break;
@@ -59,7 +60,7 @@ const run = async () => {
 
   // Pagination loop (callback)
   links = [];
-  getJson("google", { q: "Coffee" }, (page) => {
+  getJson({ engine: "google", q: "Coffee" }, (page) => {
     links.push(...extractLinks(page.organic_results));
     if (links.length < 30 && page.next) {
       page.next();