+++ tests/utils_test.ts
@@ -39,6 +39,7 @@ describe("extractNextParameters", () => {
         },
       }),
       {
+        engine: "google",
         device: "desktop",
         gl: "us",
         google_domain: "google.com",
@@ -61,6 +62,7 @@ describe("extractNextParameters", () => {
         },
       ),
       {
+        engine: "google_scholar_profiles",
         after_author: "rZlDAYoq__8J",
         hl: "en",
         mauthors: "Mike",